+++ gvm/protocols/gmpv214/types.py
@@ -23,29 +23,21 @@ from gvm.errors import InvalidArgument
 
 from gvm.protocols.gmpv208.types import (
     AggregateStatistic,
-    CredentialFormat,
-    CredentialType,
     EntityType,
     FeedType,
     FilterType,
     HostsOrdering,
     PermissionSubjectType,
-    SnmpAuthAlgorithm,
-    SnmpPrivacyAlgorithm,
     SortOrder,
     TicketStatus,
     TimeUnit,
     UserAuthType,
     get_aggregate_statistic_from_string,
-    get_credential_format_from_string,
-    get_credential_type_from_string,
     get_entity_type_from_string,
     get_feed_type_from_string,
     get_filter_type_from_string,
     get_hosts_ordering_from_string,
     get_permission_subject_type_from_string,
-    get_snmp_auth_algorithm_from_string,
-    get_snmp_privacy_algorithm_from_string,
     get_sort_order_from_string,
     get_ticket_status_from_string,
     get_time_unit_from_string,
@@ -55,40 +47,30 @@ from gvm.protocols.gmpv208.types import (
 
 __all__ = [
     "AggregateStatistic",
-    "CredentialFormat",
-    "CredentialType",
     "EntityType",
     "FeedType",
     "FilterType",
     "HostsOrdering",
     "PermissionSubjectType",
-    "ScannerType",
     "SeverityLevel",
-    "SnmpAuthAlgorithm",
-    "SnmpPrivacyAlgorithm",
     "SortOrder",
     "TicketStatus",
     "TimeUnit",
     "UserAuthType",
     "get_aggregate_statistic_from_string",
-    "get_credential_format_from_string",
-    "get_credential_type_from_string",
     "get_entity_type_from_string",
     "get_feed_type_from_string",
     "get_filter_type_from_string",
     "get_hosts_ordering_from_string",
     "get_permission_subject_type_from_string",
-    "get_scanner_type_from_string",
     "get_severity_level_from_string",
-    "get_snmp_auth_algorithm_from_string",
-    "get_snmp_privacy_algorithm_from_string",
     "get_sort_order_from_string",
     "get_ticket_status_from_string",
     "get_time_unit_from_string",
     "get_user_auth_type_from_string",
 ]
 
-
+# move this: Severity Level !!!
 class SeverityLevel(Enum):
     """Enum for severity levels"""
 
@@ -113,54 +95,3 @@ def get_severity_level_from_string(
             argument='severity_level',
             function=get_severity_level_from_string.__name__,
         ) from None
-
-
-class ScannerType(Enum):
-    """Enum for scanner type"""
-
-    OSP_SCANNER_TYPE = "1"
-    OPENVAS_SCANNER_TYPE = "2"
-    CVE_SCANNER_TYPE = "3"
-    GREENBONE_SENSOR_SCANNER_TYPE = "5"
-
-
-def get_scanner_type_from_string(
-    scanner_type: Optional[str],
-) -> Optional[ScannerType]:
-    """Convert a scanner type string to an actual ScannerType instance
-
-    Arguments:
-        scanner_type: Scanner type string to convert to a ScannerType
-    """
-    if not scanner_type:
-        return None
-
-    scanner_type = scanner_type.lower()
-
-    if (
-        scanner_type == ScannerType.OSP_SCANNER_TYPE.value
-        or scanner_type == 'osp'
-    ):
-        return ScannerType.OSP_SCANNER_TYPE
-
-    if (
-        scanner_type == ScannerType.OPENVAS_SCANNER_TYPE.value
-        or scanner_type == 'openvas'
-    ):
-        return ScannerType.OPENVAS_SCANNER_TYPE
-
-    if (
-        scanner_type == ScannerType.CVE_SCANNER_TYPE.value
-        or scanner_type == 'cve'
-    ):
-        return ScannerType.CVE_SCANNER_TYPE
-
-    if (
-        scanner_type == ScannerType.GREENBONE_SENSOR_SCANNER_TYPE.value
-        or scanner_type == 'greenbone'
-    ):
-        return ScannerType.GREENBONE_SENSOR_SCANNER_TYPE
-
-    raise InvalidArgument(
-        argument='scanner_type', function=get_scanner_type_from_string.__name__
-    )