+++ tests/protocols/gmpv208/entities/credentials/test_modify_credential.py
@@ -20,7 +20,7 @@ from gvm.errors import RequiredArgument, InvalidArgumentType
 from gvm.protocols.gmpv208 import SnmpAuthAlgorithm, SnmpPrivacyAlgorithm
 
 
-class GmpModifyCredentialTestCase:
+class GmpModifyCredentialTestMixin:
     def test_modify_credential(self):
         self.gmp.modify_credential(credential_id='c1')