+++ tests/protocols/gmpv208/entities/scanners/test_verify_scanner.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,9 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import Gmpv214TestCase
-from .testcmds import GmpModifyUserTestCase
+from gvm.errors import GvmError
 
 
-class Gmpv214ModifyUserTestCase(GmpModifyUserTestCase, Gmpv214TestCase):
-    pass
+class GmpVerifyScannerTestMixin:
+    def test_verify(self):
+        self.gmp.verify_scanner('a1')
+
+        self.connection.send.has_been_called_with(
+            '<verify_scanner scanner_id="a1"/>'
+        )
+
+    def test_missing_id(self):
+        with self.assertRaises(GvmError):
+            self.gmp.verify_scanner(None)
+
+        with self.assertRaises(GvmError):
+            self.gmp.verify_scanner('')